Move new selector to private
