+++ packages/core-data/src/selectors.ts
@@ -113,7 +113,7 @@ type Optional< T > = T | undefined;
 /**
  * HTTP Query parameters sent with the API request to fetch the entity records.
  */
-type GetRecordsHttpQuery = Record< string, any >;
+export type GetRecordsHttpQuery = Record< string, any >;
 
 /**
  * Arguments for EntityRecord selectors.
@@ -887,57 +887,6 @@ export const getEditedEntityRecord = createSelector(
 			state.entities.records?.[ kind ]?.[ name ]?.queriedData
 				.itemIsComplete[ context ]?.[ recordId ],
 			state.entities.records?.[ kind ]?.[ name ]?.edits?.[ recordId ],
-		];
-	}
-);
-
-/**
- * Returns a list of entity records, merged with their edits.
- *
- * @param state     State tree.
- * @param kind      Entity kind.
- * @param name      Entity name.
- * @param recordIds Record IDs.
- *
- * @return The list of entity records, merged with their edits.
- */
-export const getEditedEntityRecords = createSelector(
-	< EntityRecord extends ET.EntityRecord< any > >(
-		state: State,
-		kind: string,
-		name: string,
-		recordIds: EntityRecordKey[]
-	): Array< ET.Updatable< EntityRecord > | false > => {
-		return recordIds.map( ( recordId ) =>
-			getEditedEntityRecord( state, kind, name, recordId )
-		);
-	},
-	(
-		state: State,
-		kind: string,
-		name: string,
-		recordIds: EntityRecordKey[],
-		query?: GetRecordsHttpQuery
-	) => {
-		const context = query?.context ?? 'default';
-		return [
-			state.entities.config,
-			...recordIds.map(
-				( recordId ) =>
-					state.entities.records?.[ kind ]?.[ name ]?.queriedData
-						.items[ context ]?.[ recordId ]
-			),
-			...recordIds.map(
-				( recordId ) =>
-					state.entities.records?.[ kind ]?.[ name ]?.queriedData
-						.itemIsComplete[ context ]?.[ recordId ]
-			),
-			...recordIds.map(
-				( recordId ) =>
-					state.entities.records?.[ kind ]?.[ name ]?.edits?.[
-						recordId
-					]
-			),
 		];
 	}
 );
